requirements.txt
paramiko==3.4.0
requests==2.31.0
PyYAML==6.0.1
python-dateutil==2.8.2 